update and delete function
