+++ server/src/controllers/task.controller.ts
@@ -2,7 +2,7 @@ import db from "../model";
 import format from "pg-format";
 import { Request, Response } from "express";
 import { z } from "zod";
-import { addTaskDb } from "../model/task.db";
+import { addTaskDb, deleteTaskDb, updateTaskDb } from "../model/task.db";
 
 const getAllTasks = (req: Request, res: Response) => {
     console.log("running all tasks")
@@ -97,59 +97,20 @@ const updateTask = async (req: Request, res: Response) => {
         });
     else {
         try {
+            await client.query("BEGIN");
             const data = UpdateTaskConstraint.parse(req.body);
 
-            const pastData = await client.query(
-                "SELECT * FROM task WHERE user_id = $1 AND id = $2;",
-                [userId, taskId]
-            ).then(res => res.rows[0])
-    
-            // if toggle task completed
-            if (pastData.completed != data.completed) {
-                // get the last value of task_order/completed_order
-                const { max_order, completed_order } = await client.query(
-                    "SELECT MAX(task_order) as max_order, MAX(completed_order) as max_completed FROM task WHERE user_id = $1 AND duedate = $2;",
-                    [userId, pastData.duedate]
-                ).then(res => ({
-                    max_order: res.rows[0].max_order === null ? 0 : res.rows[0].max_order + 1,
-                    completed_order: res.rows[0].max_completed === null ? 0 : res.rows[0].max_completed + 1
-                }));
-    
-                // create an update order sql string
-                const dataOrder = data.completed ? format(", task_order = NULL, completed_order = %s", completed_order) :
-                                    format(", task_order = %s, completed_order = NULL", max_order);
-    
-                const sql = format(
-                    "UPDATE task SET name = $1, completed = $2%s WHERE user_id = $3 AND id = $4;",
-                    dataOrder
-                )
-                await client.query(sql, [data.name, data.completed, userId, taskId]);
-    
-                const order_string = data.completed ? "task_order" : "completed_order";
-    
-                const sqlUpdate = format(
-                    `UPDATE task SET %s = %s - 1
-                        WHERE user_id = $1 AND 
-                              %s >= $2 AND
-                              duedate = $3 returning id, name, duedate, task_order, completed_order;`, 
-                    order_string, 
-                    order_string,
-                    order_string
-                );
-                await client.query(sqlUpdate, [userId, pastData[order_string], pastData.duedate])
-                
-            } else {
-                await client.query(
-                    "UPDATE task SET name = $1, completed = $2 WHERE user_id = $3 AND id = $4;",
-                    [data.name, data.completed, userId, taskId]
-                )
-            }
+            await updateTaskDb(client, userId, taskId, data);
+
+            await client.query("COMMIT");
             res.status(200).json({ msg: "task updated" })
             
         } catch (e) {
             client.query("ROLLBACK");
             console.log(e);
             res.status(500).json({ msg: "db error" })
+        } finally {
+            client.release();
         }
     }
 }
@@ -257,25 +218,13 @@ const deleteTask = async (req: Request, res: Response) => {
         try {
             await client.query("BEGIN");
 
-            const removedTask = await client.query("DELETE FROM task WHERE id = $1 AND user_id = $2 returning *;", [taskId, userId])
-                .then(res => res.rows[0]);
-
-            const order_string = removedTask.completed ? "completed_order" : "task_order"
-            const sql = format(
-                `UPDATE task SET %s = %s - 1 
-                    WHERE user_id = $1 AND 
-                        %s >= $2 AND
-                        duedate = $3;`, 
-                order_string, 
-                order_string,
-                order_string
-            );
-            await client.query(sql, [userId, removedTask.completed ? removedTask.completed_order : removedTask.task_order, removedTask.duedate])
-
+            const removedTask = await deleteTaskDb(client, userId, taskId);
             await client.query("COMMIT");
             res.status(200).json({ msg: "task deleted", task: removedTask })
         } catch (e) {
             res.status(500).json({ msg: "db error" })
+        } finally {
+            client.release()
         }
     }
 }
